+++ Source/DataSources/CzmlDataSource.js
@@ -26,6 +26,7 @@ define([
         '../Core/LinearApproximation',
         '../Core/Math',
         '../Core/NearFarScalar',
+        '../Core/PolygonHierarchy',
         '../Core/Quaternion',
         '../Core/Rectangle',
         '../Core/ReferenceFrame',
@@ -75,6 +76,7 @@ define([
         './PolylineGraphics',
         './PolylineOutlineMaterialProperty',
         './PositionPropertyArray',
+        './Property',
         './PropertyArray',
         './PropertyBag',
         './RectangleGraphics',
@@ -117,6 +119,7 @@ define([
         LinearApproximation,
         CesiumMath,
         NearFarScalar,
+        PolygonHierarchy,
         Quaternion,
         Rectangle,
         ReferenceFrame,
@@ -166,6 +169,7 @@ define([
         PolylineGraphics,
         PolylineOutlineMaterialProperty,
         PositionPropertyArray,
+        Property,
         PropertyArray,
         PropertyBag,
         RectangleGraphics,
@@ -633,10 +637,11 @@ define([
 
     function updateInterpolationSettings(packetData, property) {
         var interpolationAlgorithm = packetData.interpolationAlgorithm;
-        if (defined(interpolationAlgorithm) || defined(packetData.interpolationDegree)) {
+        var interpolationDegree = packetData.interpolationDegree;
+        if (defined(interpolationAlgorithm) || defined(interpolationDegree)) {
             property.setInterpolationOptions({
                 interpolationAlgorithm : interpolators[interpolationAlgorithm],
-                interpolationDegree : packetData.interpolationDegree
+                interpolationDegree : interpolationDegree
             });
         }
 
@@ -665,17 +670,42 @@ define([
         iso8601 : undefined
     };
 
+    function intervalFromString(intervalString) {
+        if (!defined(intervalString)) {
+            return undefined;
+        }
+        iso8601Scratch.iso8601 = intervalString;
+        return TimeInterval.fromIso8601(iso8601Scratch);
+    }
+
+    function wrapPropertyInInfiniteInterval(property) {
+        var interval = Iso8601.MAXIMUM_INTERVAL.clone();
+        interval.data = property;
+        return interval;
+    }
+
+    function convertPropertyToComposite(property) {
+        // Create the composite and add the old property, wrapped in an infinite interval.
+        var composite = new CompositeProperty();
+        composite.intervals.addInterval(wrapPropertyInInfiniteInterval(property));
+        return composite;
+    }
+
+    function convertPositionPropertyToComposite(property) {
+        // Create the composite and add the old property, wrapped in an infinite interval.
+        var composite = new CompositePositionProperty(property.referenceFrame);
+        composite.intervals.addInterval(wrapPropertyInInfiniteInterval(property));
+        return composite;
+    }
+
     function processProperty(type, object, propertyName, packetData, constrainedInterval, sourceUri, entityCollection) {
-        var combinedInterval;
-        var packetInterval = packetData.interval;
-        if (defined(packetInterval)) {
-            iso8601Scratch.iso8601 = packetInterval;
-            combinedInterval = TimeInterval.fromIso8601(iso8601Scratch);
-            if (defined(constrainedInterval)) {
+        var combinedInterval = intervalFromString(packetData.interval);
+        if (defined(constrainedInterval)) {
+            if (defined(combinedInterval)) {
                 combinedInterval = TimeInterval.intersect(combinedInterval, constrainedInterval, scratchTimeInterval);
+            } else {
+                combinedInterval = constrainedInterval;
             }
-        } else if (defined(constrainedInterval)) {
-            combinedInterval = constrainedInterval;
         }
 
         var packedLength;
@@ -706,6 +736,10 @@ define([
 
         if (isValue) {
             unwrappedInterval = unwrapInterval(type, packetData, sourceUri);
+            if (!defined(unwrappedInterval)) {
+                // not a known value type, bail
+                return;
+            }
             packedLength = defaultValue(type.packedLength, 1);
             unwrappedIntervalLength = defaultValue(unwrappedInterval.length, 1);
             isSampled = !defined(packetData.array) && (typeof unwrappedInterval !== 'string') && (unwrappedIntervalLength > packedLength) && (type !== Object);
@@ -737,8 +771,7 @@ define([
         // replaces any non-sampled property that may exist.
         if (isSampled && !hasInterval) {
             if (!(property instanceof SampledProperty)) {
-                property = new SampledProperty(type);
-                object[propertyName] = property;
+                object[propertyName] = property = new SampledProperty(type);
             }
             property.addSamplesPackedArray(unwrappedInterval, epoch);
             updateInterpolationSettings(packetData, property);
@@ -761,16 +794,11 @@ define([
 
             // If no property exists, simply use a new interval collection
             if (!defined(property)) {
-                if (isValue) {
-                    property = new TimeIntervalCollectionProperty();
-                } else {
-                    property = new CompositeProperty();
-                }
-                object[propertyName] = property;
+                object[propertyName] = property = isValue ? new TimeIntervalCollectionProperty() : new CompositeProperty();
             }
 
             if (isValue && property instanceof TimeIntervalCollectionProperty) {
-                // If we create a collection, or it already existed, use it.
+                // If we created a collection, or it already was one, use it.
                 property.intervals.addInterval(combinedInterval);
             } else if (property instanceof CompositeProperty) {
                 // If the collection was already a CompositeProperty, use it.
@@ -780,16 +808,7 @@ define([
                 property.intervals.addInterval(combinedInterval);
             } else {
                 // Otherwise, create a CompositeProperty but preserve the existing data.
-                // Put the old property in an infinite interval.
-                interval = Iso8601.MAXIMUM_INTERVAL.clone();
-                interval.data = property;
-
-                // Create the composite.
-                property = new CompositeProperty();
-                object[propertyName] = property;
-
-                // Add the old property interval.
-                property.intervals.addInterval(interval);
+                object[propertyName] = property = convertPropertyToComposite(property);
 
                 // Change the new data to a ConstantProperty and add it.
                 if (isValue) {
@@ -801,24 +820,14 @@ define([
             return;
         }
 
-        //isSampled && hasInterval
+        // isSampled && hasInterval
         if (!defined(property)) {
-            property = new CompositeProperty();
-            object[propertyName] = property;
+            object[propertyName] = property = new CompositeProperty();
         }
 
         // Create a CompositeProperty but preserve the existing data.
         if (!(property instanceof CompositeProperty)) {
-            // Put the old property in an infinite interval.
-            interval = Iso8601.MAXIMUM_INTERVAL.clone();
-            interval.data = property;
-
-            // Create the composite.
-            property = new CompositeProperty();
-            object[propertyName] = property;
-
-            // Add the old property interval.
-            property.intervals.addInterval(interval);
+            object[propertyName] = property = convertPropertyToComposite(property);
         }
 
         // Check if the interval already exists in the composite.
@@ -862,7 +871,7 @@ define([
         }
 
         if (isArray(packetData)) {
-            for (var i = 0, len = packetData.length; i < len; i++) {
+            for (var i = 0, len = packetData.length; i < len; ++i) {
                 processProperty(type, object, propertyName, packetData[i], interval, sourceUri, entityCollection);
             }
         } else {
@@ -871,16 +880,13 @@ define([
     }
 
     function processPositionProperty(object, propertyName, packetData, constrainedInterval, sourceUri, entityCollection) {
-        var combinedInterval;
-        var packetInterval = packetData.interval;
-        if (defined(packetInterval)) {
-            iso8601Scratch.iso8601 = packetInterval;
-            combinedInterval = TimeInterval.fromIso8601(iso8601Scratch);
-            if (defined(constrainedInterval)) {
+        var combinedInterval = intervalFromString(packetData.interval);
+        if (defined(constrainedInterval)) {
+            if (defined(combinedInterval)) {
                 combinedInterval = TimeInterval.intersect(combinedInterval, constrainedInterval, scratchTimeInterval);
+            } else {
+                combinedInterval = constrainedInterval;
             }
-        } else if (defined(constrainedInterval)) {
-            combinedInterval = constrainedInterval;
         }
 
         var numberOfDerivatives = defined(packetData.cartesianVelocity) ? 1 : 0;
@@ -936,8 +942,7 @@ define([
         // replaces any non-sampled property that may exist.
         if (isSampled && !hasInterval) {
             if (!(property instanceof SampledPositionProperty) || (defined(referenceFrame) && property.referenceFrame !== referenceFrame)) {
-                property = new SampledPositionProperty(referenceFrame, numberOfDerivatives);
-                object[propertyName] = property;
+                object[propertyName] = property = new SampledPositionProperty(referenceFrame, numberOfDerivatives);
             }
             property.addSamplesPackedArray(unwrappedInterval, epoch);
             updateInterpolationSettings(packetData, property);
@@ -979,17 +984,7 @@ define([
                 property.intervals.addInterval(combinedInterval);
             } else {
                 // Otherwise, create a CompositePositionProperty but preserve the existing data.
-
-                // Put the old property in an infinite interval.
-                interval = Iso8601.MAXIMUM_INTERVAL.clone();
-                interval.data = property;
-
-                // Create the composite.
-                property = new CompositePositionProperty(property.referenceFrame);
-                object[propertyName] = property;
-
-                // Add the old property interval.
-                property.intervals.addInterval(interval);
+                object[propertyName] = property = convertPositionPropertyToComposite(property);
 
                 // Change the new data to a ConstantPositionProperty and add it.
                 if (isValue) {
@@ -1001,29 +996,19 @@ define([
             return;
         }
 
-        //isSampled && hasInterval
+        // isSampled && hasInterval
         if (!defined(property)) {
-            property = new CompositePositionProperty(referenceFrame);
-            object[propertyName] = property;
+            object[propertyName] = property = new CompositePositionProperty(referenceFrame);
         } else if (!(property instanceof CompositePositionProperty)) {
             // Create a CompositeProperty but preserve the existing data.
-            // Put the old property in an infinite interval.
-            interval = Iso8601.MAXIMUM_INTERVAL.clone();
-            interval.data = property;
-
-            // Create the composite.
-            property = new CompositePositionProperty(property.referenceFrame);
-            object[propertyName] = property;
-
-            // Add the old property interval.
-            property.intervals.addInterval(interval);
+            object[propertyName] = property = convertPositionPropertyToComposite(property);
         }
 
-        //Check if the interval already exists in the composite.
+        // Check if the interval already exists in the composite.
         var intervals = property.intervals;
         interval = intervals.findInterval(combinedInterval);
         if (!defined(interval) || !(interval.data instanceof SampledPositionProperty) || (defined(referenceFrame) && interval.data.referenceFrame !== referenceFrame)) {
-            //If not, create a SampledPositionProperty for it.
+            // If not, create a SampledPositionProperty for it.
             interval = combinedInterval.clone();
             interval.data = new SampledPositionProperty(referenceFrame, numberOfDerivatives);
             intervals.addInterval(interval);
@@ -1060,7 +1045,7 @@ define([
         }
 
         if (isArray(packetData)) {
-            for (var i = 0, len = packetData.length; i < len; i++) {
+            for (var i = 0, len = packetData.length; i < len; ++i) {
                 processPositionProperty(object, propertyName, packetData[i], interval, sourceUri, entityCollection);
             }
         } else {
@@ -1069,16 +1054,13 @@ define([
     }
 
     function processMaterialProperty(object, propertyName, packetData, constrainedInterval, sourceUri, entityCollection) {
-        var combinedInterval;
-        var packetInterval = packetData.interval;
-        if (defined(packetInterval)) {
-            iso8601Scratch.iso8601 = packetInterval;
-            combinedInterval = TimeInterval.fromIso8601(iso8601Scratch);
-            if (defined(constrainedInterval)) {
+        var combinedInterval = intervalFromString(packetData.interval);
+        if (defined(constrainedInterval)) {
+            if (defined(combinedInterval)) {
                 combinedInterval = TimeInterval.intersect(combinedInterval, constrainedInterval, scratchTimeInterval);
+            } else {
+                combinedInterval = constrainedInterval;
             }
-        } else if (defined(constrainedInterval)) {
-            combinedInterval = constrainedInterval;
         }
 
         var property = object[propertyName];
@@ -1199,7 +1181,7 @@ define([
         }
 
         if (isArray(packetData)) {
-            for (var i = 0, len = packetData.length; i < len; i++) {
+            for (var i = 0, len = packetData.length; i < len; ++i) {
                 processMaterialProperty(object, propertyName, packetData[i], interval, sourceUri, entityCollection);
             }
         } else {
@@ -1208,7 +1190,10 @@ define([
     }
 
     function processName(entity, packet, entityCollection, sourceUri) {
-        entity.name = defaultValue(packet.name, entity.name);
+        var nameData = packet.name;
+        if (defined(nameData)) {
+            entity.name = packet.name;
+        }
     }
 
     function processDescription(entity, packet, entityCollection, sourceUri) {
@@ -1245,9 +1230,10 @@ define([
             if (!defined(entity.properties)) {
                 entity.properties = new PropertyBag();
             }
-            //We cannot simply call processPacketData(entity, 'properties', propertyData, undefined, sourceUri, entityCollection)
-            //because each property of "properties" may vary separately.
-            //The properties will be accessible as entity.properties.myprop.getValue(time).
+
+            // We cannot simply call processPacketData(entity, 'properties', propertyData, undefined, sourceUri, entityCollection)
+            // because each property of "properties" may vary separately.
+            // The properties will be accessible as entity.properties.myprop.getValue(time).
 
             for (var key in propertiesData) {
                 if (propertiesData.hasOwnProperty(key)) {
@@ -1257,7 +1243,7 @@ define([
 
                     var propertyData = propertiesData[key];
                     if (isArray(propertyData)) {
-                        for (var i = 0, len = propertyData.length; i < len; i++) {
+                        for (var i = 0, len = propertyData.length; i < len; ++i) {
                             processProperty(getPropertyType(propertyData[i]), entity.properties, key, propertyData[i], undefined, sourceUri, entityCollection);
                         }
                     } else {
@@ -1268,25 +1254,36 @@ define([
         }
     }
 
+    function processReferencesArrayPacketData(object, propertyName, references, interval, entityCollection, PropertyArrayType, CompositePropertyArrayType) {
+        var properties = references.map(function(reference) {
+            return createReferenceProperty(entityCollection, reference);
+        });
+
+        if (defined(interval)) {
+            interval = intervalFromString(interval);
+            var property = object[propertyName];
+            if (!(property instanceof CompositePropertyArrayType)) {
+                // If the property was not already a CompositeProperty,
+                // create a CompositeProperty but preserve the existing data.
+
+                // Create the composite and add the old property, wrapped in an infinite interval.
+                var composite = new CompositePropertyArrayType();
+                composite.intervals.addInterval(wrapPropertyInInfiniteInterval(property));
+
+                object[propertyName] = property = composite;
+            }
+
+            interval.data = new PropertyArrayType(properties);
+            property.intervals.addInterval(interval);
+        } else {
+            object[propertyName] = new PropertyArrayType(properties);
+        }
+    }
+
     function processArrayPacketData(object, propertyName, packetData, entityCollection) {
         var references = packetData.references;
         if (defined(references)) {
-            var properties = references.map(function(reference) {
-                return createReferenceProperty(entityCollection, reference);
-            });
-
-            var iso8601Interval = packetData.interval;
-            if (defined(iso8601Interval)) {
-                iso8601Interval = TimeInterval.fromIso8601(iso8601Interval);
-                if (!(object[propertyName] instanceof CompositePositionProperty)) {
-                    iso8601Interval.data = new PropertyArray(properties);
-                    var property = new CompositeProperty();
-                    property.intervals.addInterval(iso8601Interval);
-                    object[propertyName] = property;
-                }
-            } else {
-                object[propertyName] = new PropertyArray(properties);
-            }
+            processReferencesArrayPacketData(object, propertyName, references, packetData.interval, entityCollection, PropertyArray, CompositeProperty);
         } else {
             processPacketData(Array, object, propertyName, packetData, undefined, undefined, entityCollection);
         }
@@ -1306,55 +1303,90 @@ define([
         }
     }
 
-    function processPositionsPacketData(object, propertyName, positionsData, entityCollection) {
-        if (defined(positionsData.references)) {
-            var properties = positionsData.references.map(function(reference) {
-                return createReferenceProperty(entityCollection, reference);
-            });
-
-            var iso8601Interval = positionsData.interval;
-            if (defined(iso8601Interval)) {
-                iso8601Interval = TimeInterval.fromIso8601(iso8601Interval);
-                if (!(object[propertyName] instanceof CompositePositionProperty)) {
-                    iso8601Interval.data = new PositionPropertyArray(properties);
-                    var property = new CompositePositionProperty();
-                    property.intervals.addInterval(iso8601Interval);
-                    object[propertyName] = property;
-                }
-            } else {
-                object[propertyName] = new PositionPropertyArray(properties);
-            }
+    function processPositionArrayPacketData(object, propertyName, packetData, entityCollection) {
+        var references = packetData.references;
+        if (defined(references)) {
+            processReferencesArrayPacketData(object, propertyName, references, packetData.interval, entityCollection, PositionPropertyArray, CompositePositionProperty);
         } else {
-            if (defined(positionsData.cartesian)) {
-                positionsData.array = Cartesian3.unpackArray(positionsData.cartesian);
-            } else if (defined(positionsData.cartographicRadians)) {
-                positionsData.array = Cartesian3.fromRadiansArrayHeights(positionsData.cartographicRadians);
-            } else if (defined(positionsData.cartographicDegrees)) {
-                positionsData.array = Cartesian3.fromDegreesArrayHeights(positionsData.cartographicDegrees);
+            if (defined(packetData.cartesian)) {
+                packetData.array = Cartesian3.unpackArray(packetData.cartesian);
+            } else if (defined(packetData.cartographicRadians)) {
+                packetData.array = Cartesian3.fromRadiansArrayHeights(packetData.cartographicRadians);
+            } else if (defined(packetData.cartographicDegrees)) {
+                packetData.array = Cartesian3.fromDegreesArrayHeights(packetData.cartographicDegrees);
             }
 
-            if (defined(positionsData.array)) {
-                processPacketData(Array, object, propertyName, positionsData, undefined, undefined, entityCollection);
+            if (defined(packetData.array)) {
+                processPacketData(Array, object, propertyName, packetData, undefined, undefined, entityCollection);
             }
         }
     }
 
-    function processPositions(object, propertyName, positionsData, entityCollection) {
-        if (!defined(positionsData)) {
+    function processPositionArray(object, propertyName, packetData, entityCollection) {
+        if (!defined(packetData)) {
             return;
         }
 
-        if (isArray(positionsData)) {
-            for (var i = 0, length = positionsData.length; i < length; i++) {
-                processPositionsPacketData(object, propertyName, positionsData[i], entityCollection);
+        if (isArray(packetData)) {
+            for (var i = 0, length = packetData.length; i < length; ++i) {
+                processPositionArrayPacketData(object, propertyName, packetData[i], entityCollection);
             }
         } else {
-            processPositionsPacketData(object, propertyName, positionsData, entityCollection);
+            processPositionArrayPacketData(object, propertyName, packetData, entityCollection);
+        }
+    }
+
+    function unpackCartesianArray(array) {
+        return Cartesian3.unpackArray(array);
+    }
+
+    function unpackCartographicRadiansArray(array) {
+        return Cartesian3.fromRadiansArrayHeights(array);
+    }
+
+    function unpackCartographicDegreesArray(array) {
+        return Cartesian3.fromDegreesArrayHeights(array);
+    }
+
+    function processPositionArrayOfArraysPacketData(object, propertyName, packetData, entityCollection) {
+        var references = packetData.references;
+        if (defined(references)) {
+            var properties = references.map(function(referenceArray) {
+                var tempObj = {};
+                processReferencesArrayPacketData(tempObj, 'positions', referenceArray, packetData.interval, entityCollection, PositionPropertyArray, CompositePositionProperty);
+                return tempObj.positions;
+            });
+            object[propertyName] = new PositionPropertyArray(properties);
+        } else {
+            if (defined(packetData.cartesian)) {
+                packetData.array = packetData.cartesian.map(unpackCartesianArray);
+            } else if (defined(packetData.cartographicRadians)) {
+                packetData.array = packetData.cartographicRadians.map(unpackCartographicRadiansArray);
+            } else if (defined(packetData.cartographicDegrees)) {
+                packetData.array = packetData.cartographicDegrees.map(unpackCartographicDegreesArray);
+            }
+
+            if (defined(packetData.array)) {
+                processPacketData(Array, object, propertyName, packetData, undefined, undefined, entityCollection);
+            }
+        }
+    }
+
+    function processPositionArrayOfArrays(object, propertyName, packetData, entityCollection) {
+        if (!defined(packetData)) {
+            return;
+        }
+
+        if (isArray(packetData)) {
+            for (var i = 0, length = packetData.length; i < length; ++i) {
+                processPositionArrayOfArraysPacketData(object, propertyName, packetData[i], entityCollection);
+            }
+        } else {
+            processPositionArrayOfArraysPacketData(object, propertyName, packetData, entityCollection);
         }
     }
 
     function processAvailability(entity, packet, entityCollection, sourceUri) {
-        var interval;
         var packetData = packet.availability;
         if (!defined(packetData)) {
             return;
@@ -1362,20 +1394,15 @@ define([
 
         var intervals;
         if (isArray(packetData)) {
-            var length = packetData.length;
-            for (var i = 0; i < length; i++) {
+            for (var i = 0, len = packetData.length; i < len; ++i) {
                 if (!defined(intervals)) {
                     intervals = new TimeIntervalCollection();
                 }
-                iso8601Scratch.iso8601 = packetData[i];
-                interval = TimeInterval.fromIso8601(iso8601Scratch);
-                intervals.addInterval(interval);
+                intervals.addInterval(intervalFromString(packetData[i]));
             }
         } else {
-            iso8601Scratch.iso8601 = packetData;
-            interval = TimeInterval.fromIso8601(iso8601Scratch);
             intervals = new TimeIntervalCollection();
-            intervals.addInterval(interval);
+            intervals.addInterval(intervalFromString(packetData));
         }
         entity.availability = intervals;
     }
@@ -1394,13 +1421,7 @@ define([
             return;
         }
 
-        var interval;
-        var intervalString = billboardData.interval;
-        if (defined(intervalString)) {
-            iso8601Scratch.iso8601 = intervalString;
-            interval = TimeInterval.fromIso8601(iso8601Scratch);
-        }
-
+        var interval = intervalFromString(billboardData.interval);
         var billboard = entity.billboard;
         if (!defined(billboard)) {
             entity.billboard = billboard = new BillboardGraphics();
@@ -1434,13 +1455,7 @@ define([
             return;
         }
 
-        var interval;
-        var intervalString = boxData.interval;
-        if (defined(intervalString)) {
-            iso8601Scratch.iso8601 = intervalString;
-            interval = TimeInterval.fromIso8601(iso8601Scratch);
-        }
-
+        var interval = intervalFromString(boxData.interval);
         var box = entity.box;
         if (!defined(box)) {
             entity.box = box = new BoxGraphics();
@@ -1464,20 +1479,14 @@ define([
             return;
         }
 
-        var interval;
-        var intervalString = corridorData.interval;
-        if (defined(intervalString)) {
-            iso8601Scratch.iso8601 = intervalString;
-            interval = TimeInterval.fromIso8601(iso8601Scratch);
-        }
-
+        var interval = intervalFromString(corridorData.interval);
         var corridor = entity.corridor;
         if (!defined(corridor)) {
             entity.corridor = corridor = new CorridorGraphics();
         }
 
         processPacketData(Boolean, corridor, 'show', corridorData.show, interval, sourceUri, entityCollection);
-        processPositions(corridor, 'positions', corridorData.positions, entityCollection);
+        processPositionArray(corridor, 'positions', corridorData.positions, entityCollection);
         processPacketData(Number, corridor, 'width', corridorData.width, interval, sourceUri, entityCollection);
         processPacketData(Number, corridor, 'height', corridorData.height, interval, sourceUri, entityCollection);
         processPacketData(HeightReference, corridor, 'heightReference', corridorData.heightReference, interval, sourceUri, entityCollection);
@@ -1502,13 +1511,7 @@ define([
             return;
         }
 
-        var interval;
-        var intervalString = cylinderData.interval;
-        if (defined(intervalString)) {
-            iso8601Scratch.iso8601 = intervalString;
-            interval = TimeInterval.fromIso8601(iso8601Scratch);
-        }
-
+        var interval = intervalFromString(cylinderData.interval);
         var cylinder = entity.cylinder;
         if (!defined(cylinder)) {
             entity.cylinder = cylinder = new CylinderGraphics();
@@ -1581,13 +1584,7 @@ define([
             return;
         }
 
-        var interval;
-        var intervalString = ellipseData.interval;
-        if (defined(intervalString)) {
-            iso8601Scratch.iso8601 = intervalString;
-            interval = TimeInterval.fromIso8601(iso8601Scratch);
-        }
-
+        var interval = intervalFromString(ellipseData.interval);
         var ellipse = entity.ellipse;
         if (!defined(ellipse)) {
             entity.ellipse = ellipse = new EllipseGraphics();
@@ -1621,13 +1618,7 @@ define([
             return;
         }
 
-        var interval;
-        var intervalString = ellipsoidData.interval;
-        if (defined(intervalString)) {
-            iso8601Scratch.iso8601 = intervalString;
-            interval = TimeInterval.fromIso8601(iso8601Scratch);
-        }
-
+        var interval = intervalFromString(ellipsoidData.interval);
         var ellipsoid = entity.ellipsoid;
         if (!defined(ellipsoid)) {
             entity.ellipsoid = ellipsoid = new EllipsoidGraphics();
@@ -1654,13 +1645,7 @@ define([
             return;
         }
 
-        var interval;
-        var intervalString = labelData.interval;
-        if (defined(intervalString)) {
-            iso8601Scratch.iso8601 = intervalString;
-            interval = TimeInterval.fromIso8601(iso8601Scratch);
-        }
-
+        var interval = intervalFromString(labelData.interval);
         var label = entity.label;
         if (!defined(label)) {
             entity.label = label = new LabelGraphics();
@@ -1695,13 +1680,7 @@ define([
             return;
         }
 
-        var interval;
-        var intervalString = modelData.interval;
-        if (defined(intervalString)) {
-            iso8601Scratch.iso8601 = intervalString;
-            interval = TimeInterval.fromIso8601(iso8601Scratch);
-        }
-
+        var interval = intervalFromString(modelData.interval);
         var model = entity.model;
         if (!defined(model)) {
             entity.model = model = new ModelGraphics();
@@ -1728,7 +1707,7 @@ define([
         var nodeTransformationsData = modelData.nodeTransformations;
         if (defined(nodeTransformationsData)) {
             if (isArray(nodeTransformationsData)) {
-                for (i = 0, len = nodeTransformationsData.length; i < len; i++) {
+                for (i = 0, len = nodeTransformationsData.length; i < len; ++i) {
                     processNodeTransformations(model, nodeTransformationsData[i], interval, sourceUri, entityCollection);
                 }
             } else {
@@ -1739,7 +1718,7 @@ define([
         var articulationsData = modelData.articulations;
         if (defined(articulationsData)) {
             if (isArray(articulationsData)) {
-                for (i = 0, len = articulationsData.length; i < len; i++) {
+                for (i = 0, len = articulationsData.length; i < len; ++i) {
                     processArticulations(model, articulationsData[i], interval, sourceUri, entityCollection);
                 }
             } else {
@@ -1749,29 +1728,24 @@ define([
     }
 
     function processNodeTransformations(model, nodeTransformationsData, constrainedInterval, sourceUri, entityCollection) {
-        var combinedInterval;
-        var packetInterval = nodeTransformationsData.interval;
-        if (defined(packetInterval)) {
-            iso8601Scratch.iso8601 = packetInterval;
-            combinedInterval = TimeInterval.fromIso8601(iso8601Scratch);
-            if (defined(constrainedInterval)) {
+        var combinedInterval = intervalFromString(nodeTransformationsData.interval);
+        if (defined(constrainedInterval)) {
+            if (defined(combinedInterval)) {
                 combinedInterval = TimeInterval.intersect(combinedInterval, constrainedInterval, scratchTimeInterval);
+            } else {
+                combinedInterval = constrainedInterval;
             }
-        } else if (defined(constrainedInterval)) {
-            combinedInterval = constrainedInterval;
         }
 
         var nodeTransformations = model.nodeTransformations;
         var nodeNames = Object.keys(nodeTransformationsData);
         for (var i = 0, len = nodeNames.length; i < len; ++i) {
             var nodeName = nodeNames[i];
-
             if (nodeName === 'interval') {
                 continue;
             }
 
             var nodeTransformationData = nodeTransformationsData[nodeName];
-
             if (!defined(nodeTransformationData)) {
                 continue;
             }
@@ -1796,29 +1770,24 @@ define([
     }
 
     function processArticulations(model, articulationsData, constrainedInterval, sourceUri, entityCollection) {
-        var combinedInterval;
-        var packetInterval = articulationsData.interval;
-        if (defined(packetInterval)) {
-            iso8601Scratch.iso8601 = packetInterval;
-            combinedInterval = TimeInterval.fromIso8601(iso8601Scratch);
-            if (defined(constrainedInterval)) {
+        var combinedInterval = intervalFromString(articulationsData.interval);
+        if (defined(constrainedInterval)) {
+            if (defined(combinedInterval)) {
                 combinedInterval = TimeInterval.intersect(combinedInterval, constrainedInterval, scratchTimeInterval);
+            } else {
+                combinedInterval = constrainedInterval;
             }
-        } else if (defined(constrainedInterval)) {
-            combinedInterval = constrainedInterval;
         }
 
         var articulations = model.articulations;
         var keys = Object.keys(articulationsData);
         for (var i = 0, len = keys.length; i < len; ++i) {
             var key = keys[i];
-
             if (key === 'interval') {
                 continue;
             }
 
             var articulationStageData = articulationsData[key];
-
             if (!defined(articulationStageData)) {
                 continue;
             }
@@ -1841,13 +1810,7 @@ define([
             return;
         }
 
-        var interval;
-        var intervalString = pathData.interval;
-        if (defined(intervalString)) {
-            iso8601Scratch.iso8601 = intervalString;
-            interval = TimeInterval.fromIso8601(iso8601Scratch);
-        }
-
+        var interval = intervalFromString(pathData.interval);
         var path = entity.path;
         if (!defined(path)) {
             entity.path = path = new PathGraphics();
@@ -1868,13 +1831,7 @@ define([
             return;
         }
 
-        var interval;
-        var intervalString = pointData.interval;
-        if (defined(intervalString)) {
-            iso8601Scratch.iso8601 = intervalString;
-            interval = TimeInterval.fromIso8601(iso8601Scratch);
-        }
-
+        var interval = intervalFromString(pointData.interval);
         var point = entity.point;
         if (!defined(point)) {
             entity.point = point = new PointGraphics();
@@ -1892,26 +1849,82 @@ define([
         processPacketData(Number, point, 'disableDepthTestDistance', pointData.disableDepthTestDistance, interval, sourceUri, entityCollection);
     }
 
+    function PolygonHierarchyProperty(polygon) {
+        this.polygon = polygon;
+        this._definitionChanged = new Event();
+    }
+
+    defineProperties(PolygonHierarchyProperty.prototype, {
+        isConstant : {
+            get : function() {
+                var positions = this.polygon._positions;
+                var holes = this.polygon._holes;
+                return (!defined(positions) || positions.isConstant) &&
+                       (!defined(holes) || holes.isConstant);
+            }
+        },
+        definitionChanged : {
+            get : function() {
+                return this._definitionChanged;
+            }
+        }
+    });
+
+    PolygonHierarchyProperty.prototype.getValue = function(time, result) {
+        var positions;
+        if (defined(this.polygon._positions)) {
+            positions = this.polygon._positions.getValue(time);
+        }
+
+        var holes;
+        if (defined(this.polygon._holes)) {
+            holes = this.polygon._holes.getValue(time);
+            if (defined(holes)) {
+                holes = holes.map(function(holePositions) {
+                    return new PolygonHierarchy(holePositions);
+                });
+            }
+        }
+
+        if (!defined(result)) {
+            return new PolygonHierarchy(positions, holes);
+        }
+
+        result.positions = positions;
+        result.holes = holes;
+        return result;
+    };
+
+    PolygonHierarchyProperty.prototype.equals = function(other) {
+        return this === other ||
+               (other instanceof PolygonHierarchyProperty &&
+                Property.equals(this.polygon._positions, other.polygon._positions) &&
+                Property.equals(this.polygon._holes, other.polygon._holes));
+    };
+
     function processPolygon(entity, packet, entityCollection, sourceUri) {
         var polygonData = packet.polygon;
         if (!defined(polygonData)) {
             return;
         }
 
-        var interval;
-        var intervalString = polygonData.interval;
-        if (defined(intervalString)) {
-            iso8601Scratch.iso8601 = intervalString;
-            interval = TimeInterval.fromIso8601(iso8601Scratch);
-        }
-
+        var interval = intervalFromString(polygonData.interval);
         var polygon = entity.polygon;
         if (!defined(polygon)) {
             entity.polygon = polygon = new PolygonGraphics();
         }
 
         processPacketData(Boolean, polygon, 'show', polygonData.show, interval, sourceUri, entityCollection);
-        processPositions(polygon, 'hierarchy', polygonData.positions, entityCollection);
+
+        // adapt 'position' property producing Cartesian[]
+        // and 'holes' property producing Cartesian[][]
+        // to a single property producing PolygonHierarchy
+        processPositionArray(polygon, '_positions', polygonData.positions, entityCollection);
+        processPositionArrayOfArrays(polygon, '_holes', polygonData.holes, entityCollection);
+        if (defined(polygon._positions) || defined(polygon._holes)) {
+            polygon.hierarchy = new PolygonHierarchyProperty(polygon);
+        }
+
         processPacketData(Number, polygon, 'height', polygonData.height, interval, sourceUri, entityCollection);
         processPacketData(HeightReference, polygon, 'heightReference', polygonData.heightReference, interval, sourceUri, entityCollection);
         processPacketData(Number, polygon, 'extrudedHeight', polygonData.extrudedHeight, interval, sourceUri, entityCollection);
@@ -1943,20 +1956,14 @@ define([
             return;
         }
 
-        var interval;
-        var intervalString = polylineData.interval;
-        if (defined(intervalString)) {
-            iso8601Scratch.iso8601 = intervalString;
-            interval = TimeInterval.fromIso8601(iso8601Scratch);
-        }
-
+        var interval = intervalFromString(polylineData.interval);
         var polyline = entity.polyline;
         if (!defined(polyline)) {
             entity.polyline = polyline = new PolylineGraphics();
         }
 
         processPacketData(Boolean, polyline, 'show', polylineData.show, interval, sourceUri, entityCollection);
-        processPositions(polyline, 'positions', polylineData.positions, entityCollection);
+        processPositionArray(polyline, 'positions', polylineData.positions, entityCollection);
         processPacketData(Number, polyline, 'width', polylineData.width, interval, sourceUri, entityCollection);
         processPacketData(Number, polyline, 'granularity', polylineData.granularity, interval, sourceUri, entityCollection);
         processMaterialPacketData(polyline, 'material', polylineData.material, interval, sourceUri, entityCollection);
@@ -1982,13 +1989,7 @@ define([
             return;
         }
 
-        var interval;
-        var intervalString = rectangleData.interval;
-        if (defined(intervalString)) {
-            iso8601Scratch.iso8601 = intervalString;
-            interval = TimeInterval.fromIso8601(iso8601Scratch);
-        }
-
+        var interval = intervalFromString(rectangleData.interval);
         var rectangle = entity.rectangle;
         if (!defined(rectangle)) {
             entity.rectangle = rectangle = new RectangleGraphics();
@@ -2020,20 +2021,14 @@ define([
             return;
         }
 
-        var interval;
-        var intervalString = wallData.interval;
-        if (defined(intervalString)) {
-            iso8601Scratch.iso8601 = intervalString;
-            interval = TimeInterval.fromIso8601(iso8601Scratch);
-        }
-
+        var interval = intervalFromString(wallData.interval);
         var wall = entity.wall;
         if (!defined(wall)) {
             entity.wall = wall = new WallGraphics();
         }
 
         processPacketData(Boolean, wall, 'show', wallData.show, interval, sourceUri, entityCollection);
-        processPositions(wall, 'positions', wallData.positions, entityCollection);
+        processPositionArray(wall, 'positions', wallData.positions, entityCollection);
         processArray(wall, 'minimumHeights', wallData.minimumHeights, entityCollection);
         processArray(wall, 'maximumHeights', wallData.maximumHeights, entityCollection);
         processPacketData(Number, wall, 'granularity', wallData.granularity, interval, sourceUri, entityCollection);
@@ -2115,12 +2110,13 @@ define([
             clock.clockStep = ClockStep.SYSTEM_CLOCK_MULTIPLIER;
             clock.multiplier = 1.0;
         }
-        if (defined(clockPacket.interval)) {
-            iso8601Scratch.iso8601 = clockPacket.interval;
-            var interval = TimeInterval.fromIso8601(iso8601Scratch);
+
+        var interval = intervalFromString(clockPacket.interval);
+        if (defined(interval)) {
             clock.startTime = interval.start;
             clock.stopTime = interval.stop;
         }
+
         if (defined(clockPacket.currentTime)) {
             clock.currentTime = JulianDate.fromIso8601(clockPacket.currentTime);
         }
@@ -2212,7 +2208,7 @@ define([
     }
 
     /**
-     * A {@link DataSource} which processes {@link https://github.com/AnalyticalGraphicsInc/cesium/wiki/CZML-Guide|CZML}.
+     * A {@link DataSource} which processes {@link https://github.com/AnalyticalGraphicsInc/czml-writer/wiki/CZML-Guide|CZML}.
      * @alias CzmlDataSource
      * @constructor
      *
@@ -2449,10 +2445,10 @@ define([
     CzmlDataSource.processMaterialPacketData = processMaterialPacketData;
 
     CzmlDataSource._processCzml = function(czml, entityCollection, sourceUri, updaterFunctions, dataSource) {
-        updaterFunctions = defined(updaterFunctions) ? updaterFunctions : CzmlDataSource.updaters;
+        updaterFunctions = defaultValue(updaterFunctions, CzmlDataSource.updaters);
 
         if (isArray(czml)) {
-            for (var i = 0, len = czml.length; i < len; i++) {
+            for (var i = 0, len = czml.length; i < len; ++i) {
                 processCzmlPacket(czml[i], entityCollection, updaterFunctions, sourceUri, dataSource);
             }
         } else {